// fast-bank-system-ui/src/store/types.ts
export interface UserState {
  userId: string
  token: string
  roleId: string
  role: string
  username: string
  nickname: string
  avatar: string
  phone: string
  email: string
  bankCard: string
  idCard: string
  gender: number
  workingState: number
  realName: string
  userInfoId: string
  age: number
  ctime: string
  creditStatus: number
  overdueNumber: number
}